[REFACTOR] 코드리뷰 반영_ 예외처리
